Move annotations to core module
