--- android/common/annotations/build.gradle.kts
@@ -1,28 +1,0 @@
-@file:Suppress("UnstableApiUsage")
-
-import util.libs
-
-plugins {
-    id("com.android.library")
-    kotlin("android")
-}
-
-android {
-    compileSdk = libs.versions.android.compile.get().toInt()
-    namespace = "com.thomaskioko.tvmaniac.annotations"
-
-    defaultConfig {
-        minSdk = libs.versions.android.min.get().toInt()
-        targetSdk = libs.versions.android.target.get().toInt()
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-}
-
-dependencies {
-
-    implementation(libs.inject)
-}
