--- android/common/annotations/src/main/kotlin/com/thomaskioko/tvmaniac/core/annotations/IoDispatcher.kt
@@ -1,7 +1,0 @@
-package com.thomaskioko.tvmaniac.core.annotations
-
-import javax.inject.Qualifier
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class IoDispatcher
