+++ android/features/settings/src/main/kotlin/com/thomaskioko/tvmaniac/settings/SettingsViewModel.kt
@@ -2,8 +2,8 @@ package com.thomaskioko.tvmaniac.settings
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.thomaskioko.tvmaniac.core.annotations.DefaultDispatcher
 import com.thomaskioko.tvmaniac.shared.core.ui.Store
+import com.thomaskioko.tvmaniac.shared.core.ui.di.DefaultDispatcher
 import com.thomaskioko.tvmaniac.shared.persistance.SettingsActions
 import com.thomaskioko.tvmaniac.shared.persistance.SettingsEffect
 import com.thomaskioko.tvmaniac.shared.persistance.SettingsState
