+++ shared/core/ui/src/androidMain/kotlin/com/thomaskioko/tvmaniac/shared/core/ui/di/MainDispatcher.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.shared.core.ui.di
+
+import javax.inject.Qualifier
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class MainDispatcher
