+++ shared/domain/discover/implementation/src/androidMain/kotlin/com/thomaskioko/tvmaniac/discover/implementation/di/DiscoverModule.kt
@@ -1,6 +1,5 @@
 package com.thomaskioko.tvmaniac.discover.implementation.di
 
-import com.thomaskioko.tvmaniac.core.annotations.DefaultDispatcher
 import com.thomaskioko.tvmaniac.datasource.cache.TvManiacDatabase
 import com.thomaskioko.tvmaniac.discover.api.ObserveDiscoverShowsInteractor
 import com.thomaskioko.tvmaniac.discover.api.cache.CategoryCache
@@ -10,6 +9,7 @@ import com.thomaskioko.tvmaniac.discover.implementation.CategoryCacheImpl
 import com.thomaskioko.tvmaniac.discover.implementation.DiscoverCategoryCacheImpl
 import com.thomaskioko.tvmaniac.discover.implementation.DiscoverRepositoryImpl
 import com.thomaskioko.tvmaniac.remote.api.TvShowsService
+import com.thomaskioko.tvmaniac.shared.core.ui.di.DefaultDispatcher
 import com.thomaskioko.tvmaniac.showcommon.api.cache.TvShowCache
 import dagger.Module
 import dagger.Provides
