+++ shared/domain/episodes/implementation/src/androidMain/kotlin/com/thomaskioko/tvmaniac/shared/domain/episodes/implementation/di/EpisodeModule.kt
@@ -1,6 +1,5 @@
 package com.thomaskioko.tvmaniac.shared.domain.episodes.implementation.di
 
-import com.thomaskioko.tvmaniac.core.annotations.DefaultDispatcher
 import com.thomaskioko.tvmaniac.datasource.cache.TvManiacDatabase
 import com.thomaskioko.tvmaniac.episodes.api.EpisodeRepository
 import com.thomaskioko.tvmaniac.episodes.api.EpisodesCache
@@ -8,6 +7,7 @@ import com.thomaskioko.tvmaniac.episodes.api.ObserveEpisodesInteractor
 import com.thomaskioko.tvmaniac.episodes.implementation.EpisodeRepositoryImpl
 import com.thomaskioko.tvmaniac.episodes.implementation.EpisodesCacheImpl
 import com.thomaskioko.tvmaniac.remote.api.TvShowsService
+import com.thomaskioko.tvmaniac.shared.core.ui.di.DefaultDispatcher
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
