+++ shared/domain/genre/implementation/src/androidMain/kotlin/com/thomaskioko/tvmaniac/shared/domain/genre/implementation/di/GenreModule.kt
@@ -1,6 +1,5 @@
 package com.thomaskioko.tvmaniac.shared.domain.genre.implementation.di
 
-import com.thomaskioko.tvmaniac.core.annotations.DefaultDispatcher
 import com.thomaskioko.tvmaniac.datasource.cache.TvManiacDatabase
 import com.thomaskioko.tvmaniac.genre.api.GenreCache
 import com.thomaskioko.tvmaniac.genre.api.GenreRepository
@@ -8,6 +7,7 @@ import com.thomaskioko.tvmaniac.genre.api.GetGenresInteractor
 import com.thomaskioko.tvmaniac.genre.implementation.GenreCacheImpl
 import com.thomaskioko.tvmaniac.genre.implementation.GenreRepositoryImpl
 import com.thomaskioko.tvmaniac.remote.api.TvShowsService
+import com.thomaskioko.tvmaniac.shared.core.ui.di.DefaultDispatcher
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
